Move the upgrade/install of the marker to an InstallStep
