+++ src/Umbraco.Web/Install/InstallStepCollection.cs
@@ -21,6 +21,7 @@ namespace Umbraco.Web.Install
                 a.OfType<NewInstallStep>().First(),
                 a.OfType<UpgradeStep>().First(),
                 a.OfType<FilePermissionsStep>().First(),
+                a.OfType<TelemetryIdentifierStep>().First(),
                 a.OfType<ConfigureMachineKey>().First(),
                 a.OfType<DatabaseConfigureStep>().First(),
                 a.OfType<DatabaseInstallStep>().First(),
