+++ src/Umbraco.Web/Umbraco.Web.csproj
@@ -173,6 +173,7 @@
     <Compile Include="Editors\TinyMceController.cs" />
     <Compile Include="HealthCheck\Checks\Data\DatabaseIntegrityCheck.cs" />
     <Compile Include="ImageCropperTemplateCoreExtensions.cs" />
+    <Compile Include="Install\InstallSteps\TelemetryIdentifierStep.cs" />
     <Compile Include="IUmbracoContextFactory.cs" />
     <Compile Include="Install\ChangesMonitor.cs" />
     <Compile Include="Logging\WebProfiler.cs" />
@@ -296,10 +297,8 @@
     <Compile Include="Models\LinkType.cs" />
     <Compile Include="Models\TemplateQuery\OperatorFactory.cs" />
     <Compile Include="Telemetry\ReportSiteTask.cs" />
-    <Compile Include="Telemetry\TelemetryMarkerComponent.cs" />
     <Compile Include="Telemetry\TelemetryComponent.cs" />
     <Compile Include="Telemetry\TelemetryComposer.cs" />
-    <Compile Include="Telemetry\TelemetryMarkerComposer.cs" />
     <Compile Include="Templates\HtmlLocalLinkParser.cs" />
     <Compile Include="Templates\HtmlImageSourceParser.cs" />
     <Compile Include="Templates\HtmlUrlParser.cs" />
